# improve cartesian charts performance